requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
textblob==0.17.1
sentence-transformers==2.2.2
networkx==3.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 